Move version string parsing to mullvad-types
